reorganise under connectivity, separate plotting scripts and utils
